[deploy] 기본 기능 완성

+++ demo/demo/src/main/java/buildup/server/member/service/ProfileService.java
@@ -2,6 +2,11 @@ package buildup.server.member.service;
 
 import buildup.server.activity.domain.Activity;
 import buildup.server.activity.repository.ActivityRepository;
+import buildup.server.auth.domain.MemberRefreshToken;
+import buildup.server.auth.dto.TokenRequestDto;
+import buildup.server.auth.exception.AuthErrorCode;
+import buildup.server.auth.exception.AuthException;
+import buildup.server.auth.repository.RefreshTokenRepository;
 import buildup.server.entity.Interest;
 import buildup.server.member.domain.Member;
 import buildup.server.member.domain.Profile;
@@ -21,10 +26,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
+import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -37,6 +39,7 @@ public class ProfileService {
     private final ProfileRepository profileRepository;
     private final InterestRepository interestRepository;
     private final ActivityRepository activityRepository;
+    private final RefreshTokenRepository refreshTokenRepository;
     private final S3Service s3Service;
 
     @Transactional
@@ -73,20 +76,27 @@ public class ProfileService {
     }
 
     @Transactional
-    public void updateProfileImage( MultipartFile img) {
+    public void updateProfileImage(MultipartFile img) {
         Member member = findCurrentMember();
         Profile profile = profileRepository.findById(member.getId()).get();
 
         String imgUrl = profile.getImgUrl();
 
+        if (img==null)
+            throw new MemberException(MemberErrorCode.MEMBER_PROFILE_BAD_REQUEST);
+
+        // 입력이 있으면 업로드
         if (! img.isEmpty()) {
-            // 일단 입력이 있으면 업로드. 기존 이미지 있어도 overwrite
-            String url = s3Service.uploadProfile(profile, member.getId(), img);
+            //기존 이미지 있으면 delete
+            if (imgUrl != null)
+                s3Service.deleteProfile(imgUrl);
+            String url = s3Service.uploadProfile(member.getId(), img);
             profile.setImgUrl(url);
-        } else if (imgUrl!=null) {
-            // 입력이 없는데 기존 이미지가 있었던 경우 -> 이미지 삭제
-            s3Service.deleteProfile(imgUrl);
-            profile.setImgUrl(null);
+        } else { // 입력 없으면 기존 값 삭제
+            if (imgUrl != null) {
+                s3Service.deleteProfile(imgUrl);
+                profile.setImgUrl(null);
+            }
         }
     }
 
@@ -127,7 +137,19 @@ public class ProfileService {
     // TODO: 로그인한 사용자
     private Member findCurrentMember() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        Member user = memberRepository.findByUsername(authentication.getName()).get();
+        Member user = memberRepository.findByUsername(authentication.getName())
+                .orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
         return user;
     }
+
+    @Transactional
+    public void deleteRefreshToken(TokenRequestDto tokenRequestDto) {
+        String memberId = findCurrentMember().getUsername();
+        MemberRefreshToken memberRefreshToken = refreshTokenRepository.findByUsernameAndRefreshToken(memberId, tokenRequestDto.getRefreshToken());
+        if (memberRefreshToken == null) {
+            throw new AuthException(AuthErrorCode.INVALID_REFRESH_TOKEN, "가입되지 않은 회원이거나 유효하지 않은 리프레시 토큰입니다.");
+        }
+        refreshTokenRepository.deleteById(memberRefreshToken.getRefreshTokenId());
+    }
+
 }
